backend/masonai/pyproject.toml
[project]
name = "masonai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "flask>=3.1.3",
    "flask-cors>=6.0.2",
    "openrouter>=0.7.11",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "supabase>=2.28.3",
]
